requirements.txt
Django==4.1.2
Pillow==9.2.0
django-bootstrap5==22.1
django-bootstrap3==22.1
djangorestframework==3.14.0
